backend/requirements.txt
Django
psycopg2-binary
djangorestframework
djangorestframework_simplejwt
django-cors-headers
black
flake8
pytest
pytest-django
pytest-cov
